Factor executable code generation into submodule.
This is to eventually enable a clean implementation of #1179.
