+++ src/Futhark/CodeGen/Backends/GenericC.hs
@@ -5,7 +5,6 @@
 {-# LANGUAGE TemplateHaskell #-}
 {-# LANGUAGE Trustworthy #-}
 {-# LANGUAGE TupleSections #-}
-{-# OPTIONS_GHC -fno-warn-orphans #-}
 
 -- | C code generator framework.
 module Futhark.CodeGen.Backends.GenericC
@@ -82,23 +81,19 @@ where
 import Control.Monad.Identity
 import Control.Monad.RWS
 import Data.Bifunctor (first)
-import Data.Bits (shiftR, xor)
-import Data.Char (isAlphaNum, isDigit, ord)
 import qualified Data.DList as DL
 import Data.FileEmbed
-import Data.List (unzip4)
 import Data.Loc
 import qualified Data.Map.Strict as M
 import Data.Maybe
+import Futhark.CodeGen.Backends.GenericC.CLI
 import Futhark.CodeGen.Backends.GenericC.Options
 import Futhark.CodeGen.Backends.SimpleRep
 import Futhark.CodeGen.ImpCode
 import Futhark.IR.Prop (isBuiltInFunction)
 import Futhark.MonadFreshNames
-import Futhark.Util (zEncodeString)
 import qualified Language.C.Quote.OpenCL as C
 import qualified Language.C.Syntax as C
-import Text.Printf
 
 data CompilerState s = CompilerState
   { compArrayStructs :: [((C.Type, Int), (C.Type, [C.Definition]))],
@@ -441,36 +436,6 @@ fatMemory _ = asks envFatMemory
 cacheMem :: C.ToExp a => a -> CompilerM op s (Maybe VName)
 cacheMem a = asks $ M.lookup (C.toExp a noLoc) . envCachedMem
 
-instance C.ToIdent Name where
-  toIdent = C.toIdent . zEncodeString . nameToString
-
-instance C.ToIdent VName where
-  toIdent = C.toIdent . zEncodeString . pretty
-
-instance C.ToExp VName where
-  toExp v _ = [C.cexp|$id:v|]
-
-instance C.ToExp IntValue where
-  toExp (Int8Value v) = C.toExp v
-  toExp (Int16Value v) = C.toExp v
-  toExp (Int32Value v) = C.toExp v
-  toExp (Int64Value v) = C.toExp v
-
-instance C.ToExp FloatValue where
-  toExp (Float32Value v) = C.toExp v
-  toExp (Float64Value v) = C.toExp v
-
-instance C.ToExp PrimValue where
-  toExp (IntValue v) = C.toExp v
-  toExp (FloatValue v) = C.toExp v
-  toExp (BoolValue True) = C.toExp (1 :: Int8)
-  toExp (BoolValue False) = C.toExp (0 :: Int8)
-  toExp Checked = C.toExp (1 :: Int8)
-
-instance C.ToExp SubExp where
-  toExp (Var v) = C.toExp v
-  toExp (Constant c) = C.toExp c
-
 -- | Construct a publicly visible definition using the specified name
 -- as the template.  The first returned definition is put in the
 -- header file, and the second is the implementation.  Returns the public
@@ -529,9 +494,6 @@ stms = mapM_ stm
 
 decl :: C.InitGroup -> CompilerM op s ()
 decl x = item [C.citem|$decl:x;|]
-
-addrOf :: C.Exp -> C.Exp
-addrOf e = [C.cexp|&$exp:e|]
 
 -- | Public names must have a consitent prefix.
 publicName :: String -> CompilerM op s String
@@ -810,21 +772,6 @@ allocMem mem size space on_failure = do
       freeRawMem mem space mem_s
       allocRawMem mem size space [C.cexp|desc|]
 
-primTypeInfo :: PrimType -> Signedness -> C.Exp
-primTypeInfo (IntType it) t = case (it, t) of
-  (Int8, TypeUnsigned) -> [C.cexp|u8_info|]
-  (Int16, TypeUnsigned) -> [C.cexp|u16_info|]
-  (Int32, TypeUnsigned) -> [C.cexp|u32_info|]
-  (Int64, TypeUnsigned) -> [C.cexp|u64_info|]
-  (Int8, _) -> [C.cexp|i8_info|]
-  (Int16, _) -> [C.cexp|i16_info|]
-  (Int32, _) -> [C.cexp|i32_info|]
-  (Int64, _) -> [C.cexp|i64_info|]
-primTypeInfo (FloatType Float32) _ = [C.cexp|f32_info|]
-primTypeInfo (FloatType Float64) _ = [C.cexp|f64_info|]
-primTypeInfo Bool _ = [C.cexp|bool_info|]
-primTypeInfo Cert _ = [C.cexp|bool_info|]
-
 copyMemoryDefaultSpace ::
   C.Exp ->
   C.Exp ->
@@ -839,38 +786,6 @@ copyMemoryDefaultSpace destmem destidx srcmem srcidx nbytes =
                       $exp:nbytes);|]
 
 --- Entry points.
-
-arrayName :: PrimType -> Signedness -> Int -> String
-arrayName pt signed rank =
-  prettySigned (signed == TypeUnsigned) pt ++ "_" ++ show rank ++ "d"
-
-opaqueName :: String -> [ValueDesc] -> String
-opaqueName s _
-  | valid = "opaque_" ++ s
-  where
-    valid =
-      head s /= '_'
-        && not (isDigit $ head s)
-        && all ok s
-    ok c = isAlphaNum c || c == '_'
-opaqueName s vds = "opaque_" ++ hash (zipWith xor [0 ..] $ map ord (s ++ concatMap p vds))
-  where
-    p (ScalarValue pt signed _) =
-      show (pt, signed)
-    p (ArrayValue _ space pt signed dims) =
-      show (space, pt, signed, length dims)
-
-    -- FIXME: a stupid hash algorithm; may have collisions.
-    hash =
-      printf "%x" . foldl xor 0
-        . map
-          ( iter . (* 0x45d9f3b)
-              . iter
-              . (* 0x45d9f3b)
-              . iter
-              . fromIntegral
-          )
-    iter x = ((x :: Word32) `shiftR` 16) `xor` x
 
 criticalSection :: Operations op s -> [C.BlockItem] -> [C.BlockItem]
 criticalSection ops x =
@@ -1115,10 +1030,6 @@ opaqueToCType desc vds = do
       ct <- valueDescToCType vd
       return [C.csdecl|$ty:ct *$id:(tupleField i);|]
 
-externalValueToCType :: ExternalValue -> CompilerM op s C.Type
-externalValueToCType (TransparentValue vd) = valueDescToCType vd
-externalValueToCType (OpaqueValue desc vds) = opaqueToCType desc vds
-
 prepareEntryInputs :: [ExternalValue] -> CompilerM op s [C.Param]
 prepareEntryInputs = zipWithM prepare [(0 :: Int) ..]
   where
@@ -1198,8 +1109,8 @@ prepareEntryOutputs = zipWithM prepare [(0 :: Int) ..]
 onEntryPoint ::
   Name ->
   Function op ->
-  CompilerM op s (C.Definition, C.Definition, C.Initializer)
-onEntryPoint fname function@(Function _ outputs inputs _ results args) = do
+  CompilerM op s C.Definition
+onEntryPoint fname (Function _ outputs inputs _ results args) = do
   let out_args = map (\p -> [C.cexp|&$id:(paramName p)|]) outputs
       in_args = map (\p -> [C.cexp|$id:(paramName p)|]) inputs
 
@@ -1213,8 +1124,6 @@ onEntryPoint fname function@(Function _ outputs inputs _ results args) = do
     collect' $ prepareEntryInputs args
   (entry_point_output_params, pack_entry_outputs) <-
     collect' $ prepareEntryOutputs results
-
-  (cli_entry_point, cli_init) <- cliEntryPoint fname function
 
   ctx_ty <- contextType
 
@@ -1239,7 +1148,7 @@ onEntryPoint fname function@(Function _ outputs inputs _ results args) = do
   ops <- asks envOperations
 
   return
-    ( [C.cedecl|
+    [C.cedecl|
        int $id:entry_point_function_name
            ($ty:ctx_ty *ctx,
             $params:entry_point_output_params,
@@ -1250,331 +1159,13 @@ onEntryPoint fname function@(Function _ outputs inputs _ results args) = do
          $items:(criticalSection ops critical)
 
          return ret;
-       }|],
-      cli_entry_point,
-      cli_init
-    )
+       }|]
   where
     stubParam (MemParam name space) =
       declMem name space
     stubParam (ScalarParam name ty) = do
       let ty' = primTypeToCType ty
       decl [C.cdecl|$ty:ty' $id:name;|]
-
---- CLI interface
---
--- Our strategy for CLI entry points is to parse everything into
--- host memory ('DefaultSpace') and copy the result into host memory
--- after the entry point has returned.  We have some ad-hoc frobbery
--- to copy the host-level memory blocks to another memory space if
--- necessary.  This will break if the Futhark entry point uses
--- non-trivial index functions for its input or output.
---
--- The idea here is to keep the nastyness in the wrapper, whilst not
--- messing up anything else.
-
-printPrimStm :: (C.ToExp a, C.ToExp b) => a -> b -> PrimType -> Signedness -> C.Stm
-printPrimStm dest val bt ept =
-  [C.cstm|write_scalar($exp:dest, binary_output, &$exp:(primTypeInfo bt ept), &$exp:val);|]
-
--- | Return a statement printing the given external value.
-printStm :: ExternalValue -> C.Exp -> CompilerM op s C.Stm
-printStm (OpaqueValue desc _) _ =
-  return [C.cstm|printf("#<opaque %s>", $string:desc);|]
-printStm (TransparentValue (ScalarValue bt ept _)) e =
-  return $ printPrimStm [C.cexp|stdout|] e bt ept
-printStm (TransparentValue (ArrayValue _ _ bt ept shape)) e = do
-  values_array <- publicName $ "values_" ++ name
-  shape_array <- publicName $ "shape_" ++ name
-  let num_elems = cproduct [[C.cexp|$id:shape_array(ctx, $exp:e)[$int:i]|] | i <- [0 .. rank -1]]
-  return
-    [C.cstm|{
-      $ty:bt' *arr = calloc(sizeof($ty:bt'), $exp:num_elems);
-      assert(arr != NULL);
-      assert($id:values_array(ctx, $exp:e, arr) == 0);
-      write_array(stdout, binary_output, &$exp:(primTypeInfo bt ept), arr,
-                  $id:shape_array(ctx, $exp:e), $int:rank);
-      free(arr);
-    }|]
-  where
-    rank = length shape
-    bt' = primTypeToCType bt
-    name = arrayName bt ept rank
-
-readPrimStm :: C.ToExp a => a -> Int -> PrimType -> Signedness -> C.Stm
-readPrimStm place i t ept =
-  [C.cstm|if (read_scalar(&$exp:(primTypeInfo t ept),&$exp:place) != 0) {
-        futhark_panic(1, "Error when reading input #%d of type %s (errno: %s).\n",
-              $int:i,
-              $exp:(primTypeInfo t ept).type_name,
-              strerror(errno));
-      }|]
-
-readInputs :: [ExternalValue] -> CompilerM op s [(C.Stm, C.Stm, C.Stm, C.Exp)]
-readInputs = zipWithM readInput [0 ..]
-
-readInput :: Int -> ExternalValue -> CompilerM op s (C.Stm, C.Stm, C.Stm, C.Exp)
-readInput i (OpaqueValue desc _) = do
-  stm [C.cstm|futhark_panic(1, "Cannot read input #%d of type %s\n", $int:i, $string:desc);|]
-  return ([C.cstm|;|], [C.cstm|;|], [C.cstm|;|], [C.cexp|NULL|])
-readInput i (TransparentValue (ScalarValue t ept _)) = do
-  dest <- newVName "read_value"
-  item [C.citem|$ty:(primTypeToCType t) $id:dest;|]
-  stm $ readPrimStm dest i t ept
-  return ([C.cstm|;|], [C.cstm|;|], [C.cstm|;|], [C.cexp|$id:dest|])
-readInput i (TransparentValue vd@(ArrayValue _ _ t ept dims)) = do
-  dest <- newVName "read_value"
-  shape <- newVName "read_shape"
-  arr <- newVName "read_arr"
-  ty <- valueDescToCType vd
-  item [C.citem|$ty:ty *$id:dest;|]
-
-  let t' = signedPrimTypeToCType ept t
-      rank = length dims
-      name = arrayName t ept rank
-      dims_exps = [[C.cexp|$id:shape[$int:j]|] | j <- [0 .. rank -1]]
-      dims_s = concat $ replicate rank "[]"
-
-  new_array <- publicName $ "new_" ++ name
-  free_array <- publicName $ "free_" ++ name
-
-  items
-    [C.citems|
-     typename int64_t $id:shape[$int:rank];
-     $ty:t' *$id:arr = NULL;
-     errno = 0;
-     if (read_array(&$exp:(primTypeInfo t ept),
-                    (void**) &$id:arr,
-                    $id:shape,
-                    $int:(length dims))
-         != 0) {
-       futhark_panic(1, "Cannot read input #%d of type %s%s (errno: %s).\n",
-                 $int:i,
-                 $string:dims_s,
-                 $exp:(primTypeInfo t ept).type_name,
-                 strerror(errno));
-     }|]
-
-  return
-    ( [C.cstm|assert(($exp:dest = $id:new_array(ctx, $id:arr, $args:dims_exps)) != NULL);|],
-      [C.cstm|assert($id:free_array(ctx, $exp:dest) == 0);|],
-      [C.cstm|free($id:arr);|],
-      [C.cexp|$id:dest|]
-    )
-
-prepareOutputs :: [ExternalValue] -> CompilerM op s [(C.Exp, C.Stm)]
-prepareOutputs = mapM prepareResult
-  where
-    prepareResult ev = do
-      ty <- externalValueToCType ev
-      result <- newVName "result"
-
-      case ev of
-        TransparentValue ScalarValue {} -> do
-          item [C.citem|$ty:ty $id:result;|]
-          return ([C.cexp|$id:result|], [C.cstm|;|])
-        TransparentValue (ArrayValue _ _ t ept dims) -> do
-          let name = arrayName t ept $ length dims
-          free_array <- publicName $ "free_" ++ name
-          item [C.citem|$ty:ty *$id:result;|]
-          return
-            ( [C.cexp|$id:result|],
-              [C.cstm|assert($id:free_array(ctx, $exp:result) == 0);|]
-            )
-        OpaqueValue desc vds -> do
-          free_opaque <- publicName $ "free_" ++ opaqueName desc vds
-          item [C.citem|$ty:ty *$id:result;|]
-          return
-            ( [C.cexp|$id:result|],
-              [C.cstm|assert($id:free_opaque(ctx, $exp:result) == 0);|]
-            )
-
-printResult :: [(ExternalValue, C.Exp)] -> CompilerM op s [C.Stm]
-printResult vs = fmap concat $
-  forM vs $ \(v, e) -> do
-    p <- printStm v e
-    return [p, [C.cstm|printf("\n");|]]
-
-cliEntryPoint ::
-  Name ->
-  FunctionT a ->
-  CompilerM op s (C.Definition, C.Initializer)
-cliEntryPoint fname (Function _ _ _ _ results args) = do
-  ((pack_input, free_input, free_parsed, input_args), input_items) <-
-    collect' $ unzip4 <$> readInputs args
-
-  ((output_vals, free_outputs), output_decls) <-
-    collect' $ unzip <$> prepareOutputs results
-  printstms <- printResult $ zip results output_vals
-
-  ctx_ty <- contextType
-  sync_ctx <- publicName "context_sync"
-  error_ctx <- publicName "context_get_error"
-
-  let entry_point_name = nameToString fname
-      cli_entry_point_function_name = "futrts_cli_entry_" ++ entry_point_name
-  entry_point_function_name <- publicName $ "entry_" ++ entry_point_name
-
-  pause_profiling <- publicName "context_pause_profiling"
-  unpause_profiling <- publicName "context_unpause_profiling"
-
-  let run_it =
-        [C.citems|
-                  int r;
-                  // Run the program once.
-                  $stms:pack_input
-                  if ($id:sync_ctx(ctx) != 0) {
-                    futhark_panic(1, "%s", $id:error_ctx(ctx));
-                  };
-                  // Only profile last run.
-                  if (profile_run) {
-                    $id:unpause_profiling(ctx);
-                  }
-                  t_start = get_wall_time();
-                  r = $id:entry_point_function_name(ctx,
-                                                    $args:(map addrOf output_vals),
-                                                    $args:input_args);
-                  if (r != 0) {
-                    futhark_panic(1, "%s", $id:error_ctx(ctx));
-                  }
-                  if ($id:sync_ctx(ctx) != 0) {
-                    futhark_panic(1, "%s", $id:error_ctx(ctx));
-                  };
-                  if (profile_run) {
-                    $id:pause_profiling(ctx);
-                  }
-                  t_end = get_wall_time();
-                  long int elapsed_usec = t_end - t_start;
-                  if (time_runs && runtime_file != NULL) {
-                    fprintf(runtime_file, "%lld\n", (long long) elapsed_usec);
-                    fflush(runtime_file);
-                  }
-                  $stms:free_input
-                |]
-
-  return
-    ( [C.cedecl|static void $id:cli_entry_point_function_name($ty:ctx_ty *ctx) {
-    typename int64_t t_start, t_end;
-    int time_runs = 0, profile_run = 0;
-
-    // We do not want to profile all the initialisation.
-    $id:pause_profiling(ctx);
-
-    // Declare and read input.
-    set_binary_mode(stdin);
-    $items:input_items
-
-    if (end_of_input() != 0) {
-      futhark_panic(1, "Expected EOF on stdin after reading input for %s.\n", $string:(quote (pretty fname)));
-    }
-
-    $items:output_decls
-
-    // Warmup run
-    if (perform_warmup) {
-      $items:run_it
-      $stms:free_outputs
-    }
-    time_runs = 1;
-    // Proper run.
-    for (int run = 0; run < num_runs; run++) {
-      // Only profile last run.
-      profile_run = run == num_runs -1;
-      $items:run_it
-      if (run < num_runs-1) {
-        $stms:free_outputs
-      }
-    }
-
-    // Free the parsed input.
-    $stms:free_parsed
-
-    // Print the final result.
-    if (binary_output) {
-      set_binary_mode(stdout);
-    }
-    $stms:printstms
-
-    $stms:free_outputs
-  }
-                |],
-      [C.cinit|{ .name = $string:entry_point_name,
-                      .fun = $id:cli_entry_point_function_name }|]
-    )
-
-genericOptions :: [Option]
-genericOptions =
-  [ Option
-      { optionLongName = "write-runtime-to",
-        optionShortName = Just 't',
-        optionArgument = RequiredArgument "FILE",
-        optionDescription = "Print the time taken to execute the program to the indicated file, an integral number of microseconds.",
-        optionAction = set_runtime_file
-      },
-    Option
-      { optionLongName = "runs",
-        optionShortName = Just 'r',
-        optionArgument = RequiredArgument "INT",
-        optionDescription = "Perform NUM runs of the program.",
-        optionAction = set_num_runs
-      },
-    Option
-      { optionLongName = "debugging",
-        optionShortName = Just 'D',
-        optionArgument = NoArgument,
-        optionDescription = "Perform possibly expensive internal correctness checks and verbose logging.",
-        optionAction = [C.cstm|futhark_context_config_set_debugging(cfg, 1);|]
-      },
-    Option
-      { optionLongName = "log",
-        optionShortName = Just 'L',
-        optionArgument = NoArgument,
-        optionDescription = "Print various low-overhead logging information to stderr while running.",
-        optionAction = [C.cstm|futhark_context_config_set_logging(cfg, 1);|]
-      },
-    Option
-      { optionLongName = "entry-point",
-        optionShortName = Just 'e',
-        optionArgument = RequiredArgument "NAME",
-        optionDescription = "The entry point to run. Defaults to main.",
-        optionAction = [C.cstm|if (entry_point != NULL) entry_point = optarg;|]
-      },
-    Option
-      { optionLongName = "binary-output",
-        optionShortName = Just 'b',
-        optionArgument = NoArgument,
-        optionDescription = "Print the program result in the binary output format.",
-        optionAction = [C.cstm|binary_output = 1;|]
-      },
-    Option
-      { optionLongName = "help",
-        optionShortName = Just 'h',
-        optionArgument = NoArgument,
-        optionDescription = "Print help information and exit.",
-        optionAction =
-          [C.cstm|{
-                   printf("Usage: %s [OPTION]...\nOptions:\n\n%s\nFor more information, consult the Futhark User's Guide or the man pages.\n",
-                          fut_progname, option_descriptions);
-                   exit(0);
-                  }|]
-      }
-  ]
-  where
-    set_runtime_file =
-      [C.cstm|{
-          runtime_file = fopen(optarg, "w");
-          if (runtime_file == NULL) {
-            futhark_panic(1, "Cannot open %s: %s\n", optarg, strerror(errno));
-          }
-        }|]
-    set_num_runs =
-      [C.cstm|{
-          num_runs = atoi(optarg);
-          perform_warmup = 1;
-          if (num_runs <= 0) {
-            futhark_panic(1, "Need a positive number of runs, not %s\n", optarg);
-          }
-        }|]
 
 -- | The result of compilation to C is four parts, which can be put
 -- together in various ways.  The obvious way is to concatenate all of
@@ -1638,11 +1229,8 @@ compileProg ::
   m CParts
 compileProg backend ops extra header_extra spaces options prog = do
   src <- getNameSource
-  let ((prototypes, definitions, entry_points), endstate) =
+  let ((prototypes, definitions, entry_point_decls), endstate) =
         runCompilerM ops src () compileProg'
-      (entry_point_decls, cli_entry_point_decls, entry_point_inits) =
-        unzip3 entry_points
-      option_parser = generateOptionParser "parse_options" $ genericOptions ++ options
 
   let headerdefs =
         [C.cunit|
@@ -1694,101 +1282,9 @@ $esc:util_h
 $esc:timing_h
 |]
 
-  let clidefs =
-        [C.cunit|
-$esc:("#include <string.h>")
-$esc:("#include <inttypes.h>")
-$esc:("#include <errno.h>")
-$esc:("#include <ctype.h>")
-$esc:("#include <errno.h>")
-$esc:("#include <getopt.h>")
-
-$esc:values_h
-
-static int binary_output = 0;
-static typename FILE *runtime_file;
-static int perform_warmup = 0;
-static int num_runs = 1;
-// If the entry point is NULL, the program will terminate after doing initialisation and such.
-static const char *entry_point = "main";
-
-$esc:tuning_h
-
-$func:option_parser
-
-$edecls:cli_entry_point_decls
-
-typedef void entry_point_fun(struct futhark_context*);
-
-struct entry_point_entry {
-  const char *name;
-  entry_point_fun *fun;
-};
-
-int main(int argc, char** argv) {
-  fut_progname = argv[0];
-
-  struct entry_point_entry entry_points[] = {
-    $inits:entry_point_inits
-  };
-
-  struct futhark_context_config *cfg = futhark_context_config_new();
-  assert(cfg != NULL);
-
-  int parsed_options = parse_options(cfg, argc, argv);
-  argc -= parsed_options;
-  argv += parsed_options;
-
-  if (argc != 0) {
-    futhark_panic(1, "Excess non-option: %s\n", argv[0]);
-  }
-
-  struct futhark_context *ctx = futhark_context_new(cfg);
-  assert (ctx != NULL);
-
-  char* error = futhark_context_get_error(ctx);
-  if (error != NULL) {
-    futhark_panic(1, "%s", error);
-  }
-
-  if (entry_point != NULL) {
-    int num_entry_points = sizeof(entry_points) / sizeof(entry_points[0]);
-    entry_point_fun *entry_point_fun = NULL;
-    for (int i = 0; i < num_entry_points; i++) {
-      if (strcmp(entry_points[i].name, entry_point) == 0) {
-        entry_point_fun = entry_points[i].fun;
-        break;
-      }
-    }
-
-    if (entry_point_fun == NULL) {
-      fprintf(stderr, "No entry point '%s'.  Select another with --entry-point.  Options are:\n",
-                      entry_point);
-      for (int i = 0; i < num_entry_points; i++) {
-        fprintf(stderr, "%s\n", entry_points[i].name);
-      }
-      return 1;
-    }
-
-    entry_point_fun(ctx);
-
-    if (runtime_file != NULL) {
-      fclose(runtime_file);
-    }
-
-    char *report = futhark_context_report(ctx);
-    fputs(report, stderr);
-    free(report);
-  }
-
-  futhark_context_free(ctx);
-  futhark_context_config_free(cfg);
-  return 0;
-}
-                        |]
-
   let early_decls = DL.toList $ compEarlyDecls endstate
   let lib_decls = DL.toList $ compLibDecls endstate
+  let clidefs = cliDefs options $ Functions entry_funs
   let libdefs =
         [C.cunit|
 $esc:("#ifdef _MSC_VER\n#define inline __inline\n#endif")
@@ -1798,7 +1294,7 @@ $esc:("#include <ctype.h>")
 $esc:("#include <errno.h>")
 $esc:("#include <assert.h>")
 
-$esc:(header_extra)
+$esc:header_extra
 
 $esc:lock_h
 
@@ -1821,9 +1317,10 @@ $edecls:entry_point_decls
 
   return $ CParts (pretty headerdefs) (pretty utildefs) (pretty clidefs) (pretty libdefs)
   where
-    compileProg' = do
-      let Definitions consts (Functions funs) = prog
+    Definitions consts (Functions funs) = prog
+    entry_funs = filter (functionEntry . snd) funs
 
+    compileProg' = do
       (memstructs, memfuns, memreport) <- unzip3 <$> mapM defineMemorySpace spaces
 
       get_consts <- compileConstants consts
@@ -1857,10 +1354,8 @@ $edecls:entry_point_decls
         ++ cFloat64Funs
 
     util_h = $(embedStringFile "rts/c/util.h")
-    values_h = $(embedStringFile "rts/c/values.h")
     timing_h = $(embedStringFile "rts/c/timing.h")
     lock_h = $(embedStringFile "rts/c/lock.h")
-    tuning_h = $(embedStringFile "rts/c/tuning.h")
 
 commonLibFuns :: [C.BlockItem] -> CompilerM op s ()
 commonLibFuns memreport = do
@@ -2437,11 +1932,3 @@ assignmentOperator Add {} = Just $ \d e -> [C.cexp|$id:d += $exp:e|]
 assignmentOperator Sub {} = Just $ \d e -> [C.cexp|$id:d -= $exp:e|]
 assignmentOperator Mul {} = Just $ \d e -> [C.cexp|$id:d *= $exp:e|]
 assignmentOperator _ = Nothing
-
--- | Return an expression multiplying together the given expressions.
--- If an empty list is given, the expression @1@ is returned.
-cproduct :: [C.Exp] -> C.Exp
-cproduct [] = [C.cexp|1|]
-cproduct (e : es) = foldl mult e es
-  where
-    mult x y = [C.cexp|$exp:x * $exp:y|]
